new customer redirect to /customers, added alert before delete

+++ src/app/customers/customers.component.ts
@@ -34,6 +34,8 @@ export class CustomersComponent implements OnInit {
   }
 
   handleDeleteCustomer(c: Customer) {
+    let conf= confirm("Are you sure?")
+    if(!conf) return;
     this.customerService.deleteCustomer(c.id).subscribe({
       next : (value)=> {
         this.customers=this.customers.pipe(
